MBS-9825 empty suite in test apk now throws exception
