+++ subprojects/gradle/instrumentation-tests-test-fixtures/src/main/kotlin/com/avito/instrumentation/suite/dex/FakeTestSuiteLoader.kt
@@ -1,6 +1,7 @@
 package com.avito.instrumentation.suite.dex
 
 import com.avito.instrumentation.suite.dex.check.TestSignatureCheck
+import org.funktionale.tries.Try
 import java.io.File
 
 class FakeTestSuiteLoader : TestSuiteLoader {
@@ -10,5 +11,5 @@ class FakeTestSuiteLoader : TestSuiteLoader {
     override fun loadTestSuite(
         file: File,
         testSignatureCheck: TestSignatureCheck?
-    ): List<TestInApk> = result
+    ): Try<List<TestInApk>> = Try.Success(result)
 }
